requirements.txt
torch==1.0.1
tensorflow==1.15
numpy
typing
tensorboardX
matplotlib
phonemize
librosa
multiprocessing
